🐛(frontend) fix empty left panel after deleting doc
When we were deleting a parent document, the left panel
was getting empty. It was because the panel thought that
it was a child document and was trying clear
dynamically the panel.
Now, we are checking if the document is a parent or not,
if it is a parent we just redirect to the homepage.